requirements.txt
action-tutorials-py==0.20.5
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
blinker==1.4
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.18.3
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
coloredlogs==15.0.1
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.5
Deprecated==1.2.13
dill==0.3.9
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
filelock==3.16.1
flake8==4.0.1
flatbuffers==24.3.25
fonttools==4.29.1
fs==2.4.12
fsspec==2024.10.0
future==0.18.2
gast==0.5.2
genpy==2022.1
gpg==1.16.0
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
iso8601==2.1.0
jeepney==0.7.1
Jetson.GPIO==2.1.7
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
nose2==0.9.2
notify2==0.3
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
onnxruntime==1.20.0
opencv-python==4.10.0.84
osrf-pycommon==2.0.2
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
patiencediff==0.2.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
pluggy==0.13.0
ply==3.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
pythran==0.10.0
pytools==2024.1.14
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
quality-of-service-demo-py==0.20.5
requests==2.25.1
requests-toolbelt==0.9.1
roman==3.3
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosidl-cli==3.1.5
rosidl-runtime-py==0.9.3
rosnumpy==0.0.6.2
rospkg-modules==1.5.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.8.0
screeninfo==0.8.1
seaborn==0.13.2
SecretStorage==3.3.1
sensor-msgs-py==4.2.4
serial==0.0.97
six==1.16.0
snowballstemmer==2.2.0
sros2==0.10.5
sympy==1.13.1
systemd-python==234
teleop-twist-keyboard==2.4.0
tensorrt==10.3.0
tensorrt_dispatch==10.3.0
tensorrt_lean==10.3.0
tf2-ros-py==0.25.8
tf2-tools==0.25.8
toml==0.10.2
topic-monitor==0.20.5
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ultralytics==8.3.27
ultralytics-thop==2.0.10
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
vcstool==0.3.0
vcstools==0.1.42
wadllib==1.3.6
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
